chore(cpp): split stocks files
