+++ src/CMakeLists.txt
@@ -180,6 +180,8 @@ add_executable(${MMEX_EXE} WIN32 MACOSX_BUNDLE
     stockdialog.h
     stockspanel.cpp
     stockspanel.h
+    stocks_list.cpp
+    stocks_list.h
     themes.cpp
     themes.h
     transdialog.cpp
